test(background): shard context menu listener coverage
